Generate fitted simulations from a psem model

Generate simulations from a psem model incorporating either error in fitted error. Simulations explore the possible space of what a model might predict rather than an interval for use in comparison to Bayesian posteriors for non-Bayesian models. The output format and functions draw inspiration from the tidybayes::tidybayes() library and merTools::predictInterval()

# S3 method for psem
add_fitted_sims(newdata, mod, n_sims = 1000, seed = NULL, ...)

Arguments

newdata

a data.frame of new data to predict

mod

An piecewiseSEM::psem() object to simulate from.

n_sims

number of simulation samples to construct

seed

numeric, optional argument to set seed for simulations

...

Unused dots for compatibility with generic functions.

Value

A tibble::tibble() with information about simulate values.

See also

Examples

if (FALSE) {

library(piecewiseSEM)
data(keeley)
mod <- psem(
 lm(abiotic ~ distance, data=keeley),
 lm(rich ~ abiotic + hetero, data=keeley),
 lm(hetero ~ distance, data=keeley),
 data = keeley)

newdat <- data.frame(distance=c(30, 50))

new_fitted_sims <- newdat %>%
 add_fitted_sims(mod)

head(new_fitted_sims)
}